+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## v0.0.8 - In progress
+### Changed
+- Get the Docker Compose to better emulate the production environment.
 
 ## [v0.0.7](https://github.com/hubmapconsortium/portal-ui/tree/v0.0.7) - 2020-01-08
 ### Added